guard all uses of self.vpath (esp. clean)

+++ mx.py
@@ -2609,7 +2609,7 @@ class NativeProject(Project):
     def getOutput(self, replaceVar=_replaceResultsVar):
         if self.output:
             return re.sub(r'<(.+?)>', replaceVar, self.output)
-        if self.vpath:
+        if hasattr(self.subject, "vpath") and self.vpath:
             return self.get_output_root()
         return None
 
@@ -2677,7 +2677,7 @@ class NativeBuildTask(ProjectBuildTask):
 
     def clean(self, forBuild=False):
         if not forBuild:  # assume make can do incremental builds
-            if self.subject.vpath:
+            if hasattr(self.subject, "vpath") and self.subject.vpath:
                 shutil.rmtree(self.subject.getOutput())
             else:
                 run([gmake_cmd(), 'clean'], cwd=self.subject.dir)
@@ -14379,7 +14379,7 @@ def main():
         # no need to show the stack trace when the user presses CTRL-C
         abort(1, killsig=signal.SIGINT)
 
-version = VersionSpec("5.50.3")
+version = VersionSpec("5.50.4")
 
 currentUmask = None
